Unit test

+++ src/main/java/org/example/servlet/DepartmentServlet.java
@@ -1,12 +1,11 @@
 package org.example.servlet;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import org.example.repository.exception.NotFoundException;
+import org.example.exception.NotFoundException;
 import org.example.service.DepartmentService;
 import org.example.service.impl.DepartmentServiceImpl;
 import org.example.servlet.dto.DepartmentIncomingDto;
@@ -22,15 +21,30 @@ import java.util.Optional;
 
 @WebServlet(urlPatterns = {"/department/*"})
 public class DepartmentServlet extends HttpServlet {
-    private static final DepartmentService departmentService = DepartmentServiceImpl.getInstance();
+    private final transient DepartmentService departmentService = DepartmentServiceImpl.getInstance();
     private final ObjectMapper objectMapper;
 
     public DepartmentServlet() {
         this.objectMapper = new ObjectMapper();
     }
 
+    private static void setJsonHeader(HttpServletResponse resp) {
+        resp.setContentType("application/json");
+        resp.setCharacterEncoding("UTF-8");
+    }
+
+    private static String getJson(HttpServletRequest req) throws IOException {
+        StringBuilder sb = new StringBuilder();
+        BufferedReader postData = req.getReader();
+        String line;
+        while ((line = postData.readLine()) != null) {
+            sb.append(line);
+        }
+        return sb.toString();
+    }
+
     @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException {
         setJsonHeader(resp);
 
         String responseAnswer = "";
@@ -65,11 +79,12 @@ public class DepartmentServlet extends HttpServlet {
         try {
             String[] pathPart = req.getPathInfo().split("/");
             Long departmentId = Long.parseLong(pathPart[1]);
-            resp.setStatus(HttpServletResponse.SC_OK);
-
-            if ("deleteUser".equals(pathPart[2])) {
-                Long userId = Long.parseLong(pathPart[3]);
-                departmentService.deleteUserFromDepartment(departmentId, userId);
+            resp.setStatus(HttpServletResponse.SC_NO_CONTENT);
+            if (req.getPathInfo().contains("/deleteUser/")) {
+                if ("deleteUser".equals(pathPart[2])) {
+                    Long userId = Long.parseLong(pathPart[3]);
+                    departmentService.deleteUserFromDepartment(departmentId, userId);
+                }
             } else {
                 departmentService.delete(departmentId);
             }
@@ -78,7 +93,7 @@ public class DepartmentServlet extends HttpServlet {
             responseAnswer = e.getMessage();
         } catch (Exception e) {
             resp.setStatus(HttpServletResponse.SC_BAD_REQUEST);
-            responseAnswer = "Bad request.";
+            responseAnswer = "Bad request. ";
         }
         PrintWriter printWriter = resp.getWriter();
         printWriter.write(responseAnswer);
@@ -86,7 +101,7 @@ public class DepartmentServlet extends HttpServlet {
     }
 
     @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException {
         setJsonHeader(resp);
         String json = getJson(req);
 
@@ -106,7 +121,7 @@ public class DepartmentServlet extends HttpServlet {
     }
 
     @Override
-    protected void doPut(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+    protected void doPut(HttpServletRequest req, HttpServletResponse resp) throws IOException {
         setJsonHeader(resp);
         String json = getJson(req);
 
@@ -115,7 +130,7 @@ public class DepartmentServlet extends HttpServlet {
         try {
             if (req.getPathInfo().contains("/addUser/")) {
                 String[] pathPart = req.getPathInfo().split("/");
-                if ("addUser".equals(pathPart[2])) {
+                if (pathPart.length > 3 && "addUser".equals(pathPart[2])) {
                     Long departmentId = Long.parseLong(pathPart[1]);
                     resp.setStatus(HttpServletResponse.SC_OK);
                     Long userId = Long.parseLong(pathPart[3]);
@@ -136,20 +151,5 @@ public class DepartmentServlet extends HttpServlet {
         PrintWriter printWriter = resp.getWriter();
         printWriter.write(responseAnswer);
         printWriter.flush();
-    }
-
-    private static void setJsonHeader(HttpServletResponse resp) {
-        resp.setContentType("application/json");
-        resp.setCharacterEncoding("UTF-8");
-    }
-
-    private static String getJson(HttpServletRequest req) throws IOException {
-        StringBuilder sb = new StringBuilder();
-        BufferedReader postData = req.getReader();
-        String line;
-        while ((line = postData.readLine()) != null) {
-            sb.append(line);
-        }
-        return sb.toString();
     }
 }
